backend/requirements.txt
fastapi
uvicorn
python-dotenv
spacy
scikit-learn
pdfminer.six
requests
sqlalchemy
a2wsgi
python-multipart
